fix code

+++ src/pages/Register/Register.tsx
@@ -1,6 +1,5 @@
 "use client";
 import { useState } from "react";
-import Link from "next/link";
 import { FiEye, FiEyeOff } from "react-icons/fi";
 import { FcGoogle } from "react-icons/fc";
 import "@/styles/pages/_auth.scss";
@@ -8,31 +7,63 @@ import { useRouter } from "next/navigation";
 import { signInWithPopup } from "firebase/auth";
 import { auth, provider } from "@/lib/firebase";
 import { FirebaseError } from 'firebase/app';
+import { UserService } from '@/services/UserService';
+import Spinner from "@/components/LoadingSpinner";
 
-interface UserSession {
-  username: string;
-  email: string;
-  birthdate?: string;
-  gender?: string;
+interface RegisterPageProps {
+  onOpenSignin?: () => void;
 }
 
-const RegisterPage = () => {
-  const [name, setName] = useState("");
+const RegisterPage: React.FC<RegisterPageProps> = ({ onOpenSignin }) => {
   const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
-  const [confirmPassword, setConfirmPassword] = useState("");
   const [showPassword, setShowPassword] = useState(false)
   const [error, setError] = useState<string>("");
   const router = useRouter()
+  const [isLoading, setIsLoading] = useState(false);
 
-  const handleSubmit = (e: React.FormEvent) => {
+  const checkEmailExists = async (email: string) => {
+    const checkEmail = await UserService.checkEmailExists(email);
+    if (checkEmail.status == 400) {
+      setError(checkEmail.message);
+      return false;
+    }
+
+    if (checkEmail.exists) {
+      setError(checkEmail.message);
+      return false;
+    }
+
+    return true;
+  }
+
+  const handleSubmit = async (e: React.FormEvent) => {
     e.preventDefault();
-    // Save basic info to localStorage
-    localStorage.setItem('registrationData', JSON.stringify({
-      email,
-      password
-    }));
-    router.push('/onboarding');
+    setError("");
+    setIsLoading(true);
+    localStorage.removeItem('registrationData');
+
+    try {
+      // Check if email already exists
+      const emailExists = await checkEmailExists(email);
+      if (!emailExists) {
+        return;
+      }
+
+      // Save basic info to localStorage
+      localStorage.setItem('registrationData', JSON.stringify({
+        email,
+        password
+      }));
+
+      router.push('/onboarding');
+    } catch (error) {
+      console.error("Error checking email:", error);
+      setError("An error occurred while checking the email.");
+      return;
+    } finally {
+      setIsLoading(false);
+    }
   };
 
   const toggleShowPassword = () => setShowPassword(!showPassword);
@@ -40,9 +71,16 @@ const RegisterPage = () => {
   const signUpWithGoogle = async () => {
     try {
       setError("");
+      localStorage.removeItem('registrationData');
       const result = await signInWithPopup(auth, provider);
       const user = result.user;
-      
+      setIsLoading(true);
+      // Check if email from google already exists
+      const emailExists = await checkEmailExists(user.email || "");
+      if (!emailExists) {
+        return;
+      }
+
       // Save to localStorage instead of URL params
       localStorage.setItem('registrationData', JSON.stringify({
         username: user.displayName || "",
@@ -50,22 +88,31 @@ const RegisterPage = () => {
         password: "", // No password needed for Google auth
         googleAuth: true
       }));
-      
+
       router.push('/onboarding');
     } catch (error) {
       if (error instanceof FirebaseError) {
         if (error.code === 'auth/cancelled-popup-request') {
+          setIsLoading(false);
           return;
         }
         setError(error.message);
       } else {
         setError("An unexpected error occurred");
       }
+    } finally {
+      setIsLoading(false);
     }
   };
 
   return (
     <div className="auth flex flex-col justify-center items-start">
+      {isLoading && (
+        <div className="fixed inset-0 z-50 flex items-center justify-center bg-black bg-opacity-30">
+          <Spinner size={48} className="text-[#E36B00]" />
+        </div>
+      )}
+
       <div className="auth__container !max-w-[488px]">
         <div className="auth__card !py-4 !rounded-3xl">
           <h1 className="auth__title !text-xl !font-semibold">Sign up</h1>
@@ -82,7 +129,7 @@ const RegisterPage = () => {
                   placeholder="Email"
                   value={email}
                   onChange={(e) => setEmail(e.target.value)}
-                  // required
+                // required
                 />
               </div>
             </div>
@@ -99,7 +146,7 @@ const RegisterPage = () => {
                   placeholder="Password"
                   value={password}
                   onChange={(e) => setPassword(e.target.value)}
-                  // required
+                // required
                 />
                 {showPassword ? (
                   <FiEye onClick={toggleShowPassword} className="auth__input-icon" />
@@ -115,16 +162,17 @@ const RegisterPage = () => {
               <span className="font-semibold underline text=[#494D5D]">Privacy Policy</span>
             </div>
 
-            <button type="submit" className="auth__button !bg-[#E36B00] !mt-0 !rounded-xl">
+            <button disabled={isLoading} type="submit" className="auth__button !bg-[#E36B00] !mt-0 !rounded-xl">
               Continue
             </button>
             <div className="text-sm font-normal flex flex-row flex-nowrap items-center gap-2">
-              <hr className="w-full border-t border-[#494D5D]"/>
+              <hr className="w-full border-t border-[#494D5D]" />
               or
-              <hr className="w-full border-t border-[#494D5D]"/>
+              <hr className="w-full border-t border-[#494D5D]" />
             </div>
 
             <button
+              disabled={isLoading}
               type="button"
               onClick={signUpWithGoogle}
               className="!bg-transparent text-center py-3 !mt-0 !border !border-[#494D5D] !rounded-xl !text-black flex items-center justify-center"
@@ -133,12 +181,24 @@ const RegisterPage = () => {
               <span>Continue with Google</span>
             </button>
           </form>
-
+          {error && (
+            <div
+              className={`mt-4 text-center px-4 py-2 rounded-xl font-medium bg-red-100 
+                text-red-700 border border-red-300"}`}
+              dangerouslySetInnerHTML={{ __html: error }}
+            />
+          )}
           <p className="auth__footer !mt-3.5">
             Already have an account?{" "}
-            <Link href="/login" className="auth__link !text-[#494D5D]">
+            <a
+              className="auth__link !text-[#494D5D] cursor-pointer"
+              onClick={e => {
+                e.preventDefault();
+                onOpenSignin && onOpenSignin();
+              }}
+            >
               Log in
-            </Link>
+            </a>
           </p>
         </div>
       </div>
